requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
passlib
bcrypt==4.0.1
python-jose
requests
loguru
twilio
